internal: move out inner classes and rendering code from MovementManager
